moved image utils under images/ folder
